+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Changes:
 - Handle contract call RPC result directly as `Raw` (returned as hex)
 - Use `BN_*` constants as non-specified defaults in contract interfaces
 - Remove all type overrides for node/node-template (API tracks latest)
+- Contracts ABI now supports metadata version where indexes are 0-based
+- Remove support for the old-style V2 contracts from `api-contract`
 
 
 ## 4.9.1 May 6, 2021